requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.11.0
web3==5.29.1
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2==2.9.3
whitenoise==6.1.0
Pillow==9.1.1
boto3==1.24.7
django-storages==1.12.3
python-dotenv==0.20.0
hellosign-python-sdk==4.0.0

